scripts/requirements.txt
python-dateutil==2.8.*
requests==2.26.*
beautifulsoup4==4.10.*
